clean up seqr xbrowse dependencies
