+++ seqr/views/apis/analysis_group_api.py
@@ -5,14 +5,13 @@ from django.db import IntegrityError
 from django.views.decorators.csrf import csrf_exempt
 
 from seqr.models import AnalysisGroup, Family, CAN_EDIT
-from seqr.model_utils import create_seqr_model, delete_seqr_model, find_matching_xbrowse_model
+from seqr.model_utils import create_seqr_model, delete_seqr_model, update_xbrowse_family_group_families
 from seqr.views.apis.auth_api import API_LOGIN_REQUIRED_URL
 from seqr.views.utils.json_utils import create_json_response
 from seqr.views.utils.json_to_orm_utils import update_model_from_json
 from seqr.views.utils.orm_to_json_utils import get_json_for_analysis_group
 from seqr.views.utils.permissions_utils import get_project_and_check_permissions
 
-from xbrowse_server.base.models import Family as BaseFamily
 
 logger = logging.getLogger(__name__)
 
@@ -58,9 +57,7 @@ def update_analysis_group_handler(request, project_guid, analysis_group_guid=Non
                 ))
 
     analysis_group.families.set(families)
-    base_family_group = find_matching_xbrowse_model(analysis_group)
-    if base_family_group:
-        base_family_group.families.set(BaseFamily.objects.filter(seqr_family__in=families))
+    update_xbrowse_family_group_families(analysis_group, families)
 
     return create_json_response({
         'analysisGroupsByGuid': {
