+++ seqr/views/apis/locus_list_api.py
@@ -9,14 +9,14 @@ from guardian.shortcuts import assign_perm, remove_perm, get_objects_for_group
 
 from reference_data.models import GENOME_VERSION_GRCh37
 from seqr.models import LocusList, LocusListGene, LocusListInterval, IS_OWNER, CAN_VIEW, CAN_EDIT
-from seqr.model_utils import get_or_create_seqr_model, create_seqr_model, delete_seqr_model, find_matching_xbrowse_model
+from seqr.model_utils import get_or_create_seqr_model, create_seqr_model, delete_seqr_model, \
+    add_xbrowse_project_gene_lists, remove_xbrowse_project_gene_lists
 from seqr.views.apis.auth_api import API_LOGIN_REQUIRED_URL
 from seqr.utils.gene_utils import get_genes, parse_locus_list_items
 from seqr.views.utils.json_utils import create_json_response
 from seqr.views.utils.json_to_orm_utils import update_model_from_json
 from seqr.views.utils.orm_to_json_utils import get_json_for_locus_lists, get_json_for_locus_list
 from seqr.views.utils.permissions_utils import get_project_and_check_permissions, check_object_permissions, check_public_object_permissions
-from xbrowse_server.base.models import ProjectGeneList as BaseProjectGeneList
 
 logger = logging.getLogger(__name__)
 
@@ -113,14 +113,11 @@ def delete_locus_list_handler(request, locus_list_guid):
 @csrf_exempt
 def add_project_locus_lists(request, project_guid):
     project = get_project_and_check_permissions(project_guid, request.user, CAN_EDIT)
-    xbrowse_project = find_matching_xbrowse_model(project)
     request_json = json.loads(request.body)
     locus_lists = LocusList.objects.filter(guid__in=request_json['locusListGuids'])
     for locus_list in locus_lists:
         assign_perm(user_or_group=project.can_view_group, perm=CAN_VIEW, obj=locus_list)
-        xbrowse_gene_list = find_matching_xbrowse_model(locus_list)
-        if xbrowse_project and xbrowse_gene_list:
-            BaseProjectGeneList.objects.get_or_create(project=xbrowse_project, gene_list=xbrowse_gene_list)
+    add_xbrowse_project_gene_lists(project, locus_lists)
 
     return create_json_response({
         'locusListGuids': [locus_list['locusListGuid'] for locus_list in get_sorted_project_locus_lists(project, request.user)],
@@ -131,14 +128,11 @@ def add_project_locus_lists(request, project_guid):
 @csrf_exempt
 def delete_project_locus_lists(request, project_guid):
     project = get_project_and_check_permissions(project_guid, request.user, CAN_EDIT)
-    xbrowse_project = find_matching_xbrowse_model(project)
     request_json = json.loads(request.body)
     locus_lists = LocusList.objects.filter(guid__in=request_json['locusListGuids'])
     for locus_list in locus_lists:
         remove_perm(user_or_group=project.can_view_group, perm=CAN_VIEW, obj=locus_list)
-        xbrowse_gene_list = find_matching_xbrowse_model(locus_list)
-        if xbrowse_project and xbrowse_gene_list:
-            BaseProjectGeneList.objects.filter(project=xbrowse_project, gene_list=xbrowse_gene_list).delete()
+    remove_xbrowse_project_gene_lists(project, locus_lists)
 
     return create_json_response({
         'locusListGuids': [locus_list['locusListGuid'] for locus_list in get_sorted_project_locus_lists(project, request.user)],
